modularized Login
