+++ frontend/components/login/HeaderLogin.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export default function HeaderLogin() {
+    return <View></View>;
+}
