+++ frontend/utils/auth.ts
@@ -1,7 +1,4 @@
-import {
-    signInWithPhoneNumber,
-    ConfirmationResult,
-} from 'firebase/auth';
+import { signInWithPhoneNumber, ConfirmationResult } from 'firebase/auth';
 import { auth } from '@/firebase/firebaseConfig'; // Adjust path to your Firebase config
 
 // Create a variable to hold the confirmation result
